+++ service/src/services/communication-adapter/communication-adapter.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { MatrixAgentPoolModule } from '@services/matrix/agent-pool/matrix.agent.pool.module';
-import { MatrixUserManagementModule } from '@services/matrix/management/matrix.user.management.module';
 import { MatrixRoomAdapterModule } from '@services/matrix/adapter-room/matrix.room.adapter.module';
 import { MatrixAgentModule } from '@services/matrix/agent/matrix.agent.module';
 import { MatrixUserAdapterModule } from '@services/matrix/adapter-user/matrix.user.adapter.module';
 import { CommunicationAdapter } from './communication.adapter';
-import { CommunicationAdminUserModule } from '../communication-admin-user/communication.admin.user.module';
+import { MatrixAdminUserElevatedModule } from '../matrix-admin/user-elevated/matrix.admin.user.elevated.module';
+import { MatrixUserManagementModule } from '../matrix-admin/user/matrix.admin.user.module';
 
 @Module({
   imports: [
@@ -14,7 +14,7 @@ import { CommunicationAdminUserModule } from '../communication-admin-user/commun
     MatrixRoomAdapterModule,
     MatrixAgentModule,
     MatrixAgentPoolModule,
-    CommunicationAdminUserModule,
+    MatrixAdminUserElevatedModule,
   ],
   providers: [CommunicationAdapter],
   exports: [CommunicationAdapter],
