--- service/src/services/matrix-admin/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './matrix.admin.dto.event.update.room.state.for.admin.rooms';
