+++ service/src/services/matrix-admin/rooms/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './matrix.admin.roomsdto.event.update.room.state.for.admin.rooms';
