+++ service/src/services/matrix-admin/rooms/dto/matrix.admin.rooms.base.event.response.payload.ts
