+++ service/src/services/matrix-admin/rooms/dto/matrix.admin.rooms.dto.event.log.room.state.ts
