+++ service/src/services/matrix-admin/rooms/dto/matrix.admin.roomsdto.event.update.room.state.for.admin.rooms.ts
