+++ service/src/services/matrix-admin/rooms/matrix.admin.rooms.event.type.ts
@@ -1,4 +1,4 @@
-export enum MatrixAdminEventType {
+export enum MatrixAdminRoomsEventType {
   UPDATE_ROOM_STATE_FOR_ADMIN_ROOMS = 'updateRoomStateForAdminRooms',
   LOG_ROOM_STATE = 'logRoomState',
 }
