+++ service/src/services/matrix-admin/user/matrix.admin.user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { MatrixCryptographyModule } from '@services/matrix/cryptography/matrix.cryptography.module';
-import { MatrixUserManagementService } from '@services/matrix/management/matrix.user.management.service';
-import { MatrixUserAdapterModule } from '../adapter-user/matrix.user.adapter.module';
+import { MatrixUserAdapterModule } from '@src/services/matrix/adapter-user/matrix.user.adapter.module';
+import { MatrixUserManagementService } from './matrix.admin.user.service';
 
 @Module({
   imports: [MatrixUserAdapterModule, MatrixCryptographyModule, HttpModule],
