+++ service/src/services/matrix/adapter-room/dto/matrix.room.dto.options.ts
