+++ service/src/services/matrix/adapter-room/matrix.room.adapter.ts
@@ -2,7 +2,7 @@ import { RoomDirectResult } from '@alkemio/matrix-adapter-lib';
 import { RoomResult, IMessage, IReaction } from '@alkemio/matrix-adapter-lib';
 import { LogContext } from '@common/enums/index';
 import { MatrixEntityNotFoundException } from '@common/exceptions/matrix.entity.not.found.exception';
-import pkg  from '@nestjs/common';
+import pkg from '@nestjs/common';
 const { Inject, Injectable } = pkg;
 import {
   Direction,
@@ -15,21 +15,30 @@ import {
   KnownMembership,
   MatrixClient,
   MatrixEvent,
+  MsgType,
   Preset,
+  RelationType,
   TimelineWindow,
   Visibility,
 } from 'matrix-js-sdk';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { MatrixMessageAdapter } from '../adapter-message/matrix.message.adapter';
-import { MatrixRoom } from './matrix.room';
-import { MatrixRoomResponseMessage } from './matrix.room.dto.response.message';
+import { MatrixRoom } from './dto/matrix.room';
+import { MatrixRoomResponseMessage } from './dto/matrix.room.dto.response.message';
 import {
+  ReactionEventContent,
   RoomHistoryVisibilityEventContent,
   RoomJoinRulesEventContent,
+  RoomMessageEventContent,
 } from 'matrix-js-sdk/lib/types';
 import { MatrixAgent } from '../agent/matrix.agent';
-import { IRoomOpts } from './matrix.room.dto.options';
-
+import { IRoomOpts } from './dto/matrix.room.dto.options';
+import { IMatrixAgent } from '../agent/matrix.agent.interface';
+import { MatrixAgentMessageRequest } from '../agent/dto/matrix.agent.dto.message.request';
+import { MatrixAgentMessageReaction } from '../agent/dto/matrix.agent.dto.message.reaction';
+import { MatrixAgentMessageReply } from '../agent/dto/matrix.agent.dto.message.reply';
+import { MatrixAgentMessageRequestDirect } from '../agent/dto/matrix.agent.dto.message.request.direct';
+import { MatrixUserAdapter } from '../adapter-user/matrix.user.adapter';
 
 // Note: before
 @Injectable()
@@ -37,8 +46,40 @@ export class MatrixRoomAdapter {
   constructor(
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
     private readonly logger: pkg.LoggerService,
-    private matrixMessageAdapter: MatrixMessageAdapter
+    private matrixMessageAdapter: MatrixMessageAdapter,
+    private matrixUserAdapter: MatrixUserAdapter
   ) {}
+
+  async getJoinedRooms(matrixClient: MatrixClient): Promise<string[]> {
+    const response = (await matrixClient.getJoinedRooms()) as any as {
+      joined_rooms: string[];
+    };
+    return response.joined_rooms;
+  }
+
+  async isUserMemberOfRoom(
+    matrixClient: MatrixClient,
+    roomID: string
+  ): Promise<boolean> {
+    const rooms = await this.getJoinedRooms(matrixClient);
+    const roomFound = rooms.find(r => r === roomID);
+    if (roomFound) {
+      this.logger.verbose?.(
+        `[Membership] user (${matrixClient.getUserId()}) is a member of: ${roomID}`,
+        LogContext.COMMUNICATION
+      );
+      return true;
+    }
+    return false;
+  }
+
+  async logJoinedRooms(matrixClient: MatrixClient) {
+    const rooms = await this.getJoinedRooms(matrixClient);
+    this.logger.verbose?.(
+      `[Membership] user (${matrixClient.getUserId()}) rooms: ${rooms}`,
+      LogContext.COMMUNICATION
+    );
+  }
 
   async storeDirectMessageRoom(
     matrixClient: MatrixClient,
@@ -50,6 +91,225 @@ export class MatrixRoomAdapter {
 
     dmRooms[userId] = [roomId];
     await matrixClient.setAccountData(EventType.Direct, dmRooms);
+  }
+
+  async getDirectRooms(matrixAgent: IMatrixAgent): Promise<MatrixRoom[]> {
+    const matrixClient = matrixAgent.matrixClient;
+    const rooms: MatrixRoom[] = [];
+
+    // Direct rooms
+    const dmRoomMap =
+      this.getDirectMessageRoomsMap(matrixClient);
+
+    // UPDATED TO USE ASYNC ROOM ACCESS
+    for (const matrixUsername of Object.keys(dmRoomMap)) {
+      const roomId = dmRoomMap[matrixUsername][0];
+      const room = await this.getRoom(matrixAgent, roomId); // NOW ASYNC
+
+      room.receiverCommunicationsID =
+        this.matrixUserAdapter.convertMatrixUsernameToMatrixID(matrixUsername);
+      room.isDirect = true;
+      rooms.push(room);
+    }
+    return rooms;
+  }
+
+  async getRoom(
+    matrixAgent: IMatrixAgent,
+    roomId: string
+  ): Promise<MatrixRoom> {
+    // SLIDING SYNC APPROACH
+    const matrixRoom = await (matrixAgent as MatrixAgent).getRoomAsync(roomId);
+
+    if (!matrixRoom) {
+      throw new MatrixEntityNotFoundException(
+        `[User: ${matrixAgent.matrixClient.getUserId()}] Unable to access Room (${roomId}). Room either does not exist or user does not have access.`,
+        LogContext.COMMUNICATION
+      );
+    }
+
+    return matrixRoom;
+  }
+
+  async initiateMessagingToUser(
+    matrixAgent: IMatrixAgent,
+    messageRequest: MatrixAgentMessageRequestDirect
+  ): Promise<string> {
+    const directRoom = await this.getDirectRoomForMatrixID(
+      matrixAgent,
+      messageRequest.matrixID
+    );
+    if (directRoom) return directRoom.roomId;
+
+    // Room does not exist, create...
+    const targetRoomId = await this.createRoom(
+      matrixAgent.matrixClient,
+      {},
+      {
+        dmUserId: messageRequest.matrixID,
+      }
+    );
+
+    await this.storeDirectMessageRoom(
+      matrixAgent.matrixClient,
+      targetRoomId,
+      messageRequest.matrixID
+    );
+
+    return targetRoomId;
+  }
+
+  /*
+    the naming is really confusing
+    what we attempt to solve is a race condition
+    where two or more DM rooms are created between the two users
+    we aim to always resolve the
+  */
+  async getDirectUserMatrixIDForRoomID(
+    matrixAgent: MatrixAgent,
+    matrixRoomId: string
+  ): Promise<string | undefined> {
+    // Need to implement caching for performance
+    const dmRoomByUserMatrixIDMap =
+      this.getDirectMessageRoomsMap(
+        matrixAgent.matrixClient
+      );
+    const dmUserMatrixIDs = Object.keys(dmRoomByUserMatrixIDMap);
+    const dmRoom = dmUserMatrixIDs.find(
+      userID => dmRoomByUserMatrixIDMap[userID].indexOf(matrixRoomId) !== -1
+    );
+    return dmRoom;
+  }
+
+  async getDirectRoomForMatrixID(
+    matrixAgent: IMatrixAgent,
+    matrixUserId: string
+  ): Promise<MatrixRoom | undefined> {
+    const matrixUsername =
+      this.matrixUserAdapter.convertMatrixIDToUsername(matrixUserId);
+    // Need to implement caching for performance
+    const dmRoomIds = this.getDirectMessageRoomsMap(
+      matrixAgent.matrixClient
+    )[matrixUsername];
+
+    if (!dmRoomIds || !Boolean(dmRoomIds[0])) {
+      return undefined;
+    }
+
+    // Have a result
+    const targetRoomId = dmRoomIds[0];
+    return await this.getRoom(matrixAgent, targetRoomId);
+  }
+
+  async sendMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    messageRequest: MatrixAgentMessageRequest
+  ): Promise<string> {
+    const content: RoomMessageEventContent = {
+      body: messageRequest.text,
+      msgtype: MsgType.Text,
+    };
+    const response = await matrixAgent.matrixClient.sendEvent(
+      roomId,
+      EventType.RoomMessage,
+      content
+    );
+
+    return response.event_id;
+  }
+
+  async sendReplyToMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    messageRequest: MatrixAgentMessageReply
+  ): Promise<string> {
+    const content: RoomMessageEventContent = {
+      msgtype: MsgType.Text,
+      body: messageRequest.text,
+      ['m.relates_to']: {
+        rel_type: RelationType.Thread,
+        event_id: messageRequest.threadID,
+        //is_falling_back: true,
+        // when events need to be represented in an unthreaded client, this field makes the event a reply to the thread root event
+        // ['m.in_reply_to']: {
+        //   event_id: messageRequest.threadID,
+        // },
+      },
+    };
+
+    const response = await matrixAgent.matrixClient.sendEvent(
+      roomId,
+      EventType.RoomMessage,
+      content
+    );
+
+    return response.event_id;
+  }
+
+  async addReactionOnMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    messageReaction: MatrixAgentMessageReaction
+  ): Promise<string> {
+    const content: ReactionEventContent = {
+      'm.relates_to': {
+        rel_type: RelationType.Annotation,
+        event_id: messageReaction.messageID,
+        key: messageReaction.emoji,
+      },
+    };
+
+    const response = await matrixAgent.matrixClient.sendEvent(
+      roomId,
+      EventType.Reaction,
+      content
+    );
+
+    return response.event_id;
+  }
+
+  async removeReactionOnMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    reactionID: string
+  ) {
+    await matrixAgent.matrixClient.redactEvent(roomId, reactionID);
+  }
+
+  // TODO - see if the js sdk supports message aggregation
+  async editMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    messageId: string,
+    messageRequest: MatrixAgentMessageRequest
+  ) {
+    this.logger.verbose?.(
+      `Editing message: ${messageId} ${matrixAgent} ${roomId} ${messageRequest}`,
+      LogContext.COMMUNICATION
+    );
+    // const newContent: IContent = {
+    //   msgtype: MsgType.Text,
+    //   body: messageRequest.text,
+    // };
+    // const content: RoomMessageEventContent = {
+    //   msgtype: MsgType.Text,
+    //   'm.new_content': newContent,
+    //   'm.relates_to': {
+    //     rel_type: 'm.replace',
+    //     event_id: messageId,
+    //   },
+    //   body: messageRequest.text,
+    // };
+    // await matrixAgent.matrixClient.sendMessage(roomId, content);
+  }
+
+  async deleteMessage(
+    matrixAgent: IMatrixAgent,
+    roomId: string,
+    messageId: string
+  ) {
+    await matrixAgent.matrixClient.redactEvent(roomId, messageId);
   }
 
   // there could be more than one dm room per user
