+++ service/src/services/matrix/agent-pool/matrix.agent.pool.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MatrixAgentPool } from '@services/matrix/agent-pool/matrix.agent.pool';
-import { MatrixUserManagementModule } from '@services/matrix/management/matrix.user.management.module';
 import { MatrixAgentModule } from '../agent/matrix.agent.module';
+import { MatrixUserManagementModule } from '@src/services/matrix-admin/user/matrix.admin.user.module';
 
 @Module({
   imports: [MatrixUserManagementModule, MatrixAgentModule],
