+++ service/src/services/matrix/agent-pool/matrix.agent.pool.ts
@@ -3,7 +3,6 @@ import { MatrixAgentPoolException } from '@common/exceptions/matrix.agent.pool.e
 import pkg  from '@nestjs/common';
 const { Inject, Injectable } = pkg;
 import { OnModuleDestroy, OnModuleInit } from '@nestjs/common';
-import { MatrixUserManagementService } from '@services/matrix/management/matrix.user.management.service';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { MatrixAgent } from '../agent/matrix.agent';
 import { MatrixAgentService } from '../agent/matrix.agent.service';
@@ -11,6 +10,7 @@ import { ConfigService } from '@nestjs/config';
 import { Disposable } from '@src/common/interfaces/disposable.interface';
 import { NotSupportedException } from '@src/common/exceptions/not.supported.exception';
 import { LogContext } from '@src/common/enums/logging.context';
+import { MatrixUserManagementService } from '@src/services/matrix-admin/user/matrix.admin.user.service';
 
 @Injectable()
 export class MatrixAgentPool
