+++ service/src/services/matrix/events/matrix.event.dispatcher.ts
@@ -12,7 +12,7 @@ import {
   Room,
 } from 'matrix-js-sdk';
 import { first, fromEvent, Observable, Observer, Subscription } from 'rxjs';
-import { MatrixRoom } from '../adapter-room/matrix.room';
+import { MatrixRoom } from '../adapter-room/dto/matrix.room';
 import { MatrixEventHandler } from '../types/matrix.event.handler.type';
 
 export enum InternalEventNames {
